requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
config==0.5.1
cron-descriptor==1.4.3
Django==5.0.6
django-celery-beat==2.6.0
django-debug-toolbar==4.4.2
django-filter==24.2
django-js-asset==1.2.2
django-modeltranslation==0.19.3
django-mptt==0.16.0
django-session-timeout==0.1.0
django-templated-mail==1.1.1
django-timezone-field==7.0
djangorestframework==3.15.1
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
exceptiongroup==1.2.1
flower==2.0.1
GDAL==3.4.1
httptools==0.6.1
humanize==4.10.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
pytest==8.2.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rest-framework-generic-relations==2.2.0
rpds-py==0.18.1
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
